[2019.01.30] 7,8단원 예습 완료

+++ es6/chapter3/datatypeconvert.js
@@ -5,6 +5,7 @@ console.log(num);
 
 const a = parseInt("16 volts", 10);
 const b = parseInt("3a", 16);
-const c = parseInt("15.5 kph");
+const c = parseFloat("15.5 kph");
+
 
 console.log(a, b, c);
